[Silver II] Title: 예산, Time: 156 ms, Memory: 16356 KB -BaekjoonHub

+++ 백준/Silver/2512. 예산/예산.java
@@ -6,24 +6,24 @@ import java.util.Arrays;
 public class Main {
     static int N;
     static int[] arr;
-    static long total;
+    static int total;
     public static void main(String[] args) throws IOException {
         BufferedReader br = new BufferedReader(new InputStreamReader(System.in));
         N = Integer.parseInt(br.readLine());
         arr = Arrays.stream(br.readLine().split(" ")).mapToInt(Integer::parseInt).toArray();
-        total = Long.parseLong(br.readLine());
+        total = Integer.parseInt(br.readLine());
 
         Arrays.sort(arr);
-        long result = arr[N-1];
+        int result = arr[N-1];
         if (check(result)) {
             System.out.println(result);
             return;
         }
 
-        long left = 0;
-        long right = arr[N-1];
+        int left = 0;
+        int right = arr[N-1];
         while(left <= right) {
-            long mid = (left+right)/2;
+            int mid = (left+right)/2;
             if(check(mid)) {
                 result = mid;
                 left = mid + 1;
@@ -33,8 +33,8 @@ public class Main {
         System.out.println(result);
 
     }
-    static boolean check(long topPrice) {
-        long sum = 0;
+    static boolean check(int topPrice) {
+        int sum = 0;
         for(int a: arr) {
             if(a <= topPrice) sum += a;
             else sum += topPrice;
